add 'speech_from_key' dictionary
The code for working out what (if anything) to say is quite repetitive,
and that's bad.  A better way is to use a Python dictionary:

https://docs.python.org/3.4/tutorial/datastructures.html#dictionaries

(That description is quite technical but does give plenty of examples.)

We do not yet use the dictionary we've defined.  It will replace the big
chain of if/elif code shortly.