variable to track user's typing for custom speech
To allow the player to type in their own choice of speech for the robot,
we'll use a variable to keep track of the current text.  It will work
like:

Press 'y' to start typing in something of your own choice for the robot
to say.  When you press RETURN, the robot remembers what you've typed.
You'll then be able to press 'j' to make the robot say it.

Currently, nothing is done with the variable except resetting to '' (the
empty string) when the player enters 'custom speech' mode (via the 'y'
key).