add keys pressed to user speech
If we're in 'enter your own choice of speech mode', every key that gets
pressed should add a letter (or number, or punctuation, etc.) to the
variable tracking the custom speech.

http://www.pygame.org/docs/ref/key.html

tells us that the 'key pressed event' has a 'unicode' attribute saying
what letter (etc.) the key means.

For strings, '+=' means 'glue on to the end'.