variable for whether to draw snot
I noticed that the snot appears when you type spacebar in the middle of
typing in your own speech.  We need to be more careful about when we
draw the snot.  First thing to do is make a variable saying whether to
draw the snot, and then in a minute we'll set that variable more
carefully.  For now, the bug will still happen.