variable for x-coord of speech
It looks a bit odd that the speech isn't always in the middle of the
mouth.  The problem is that the 'x' coord needs to change, depending on
how long the speech is.  First step is to make this be a variable, even
though for now it's just the same number.  In a minute we'll set this
variable's value more carefully.