centre speech in mouth
Work out where to put the speech.  We can find out the width by finding
the size (width and height) of the piece of text pygame draws for us:

http://www.pygame.org/docs/ref/surface.html#pygame.Surface.get_size

Then the correct x value to start the speech is to make half of it stick
out to the left of the central x value.