keep tear y-coord in list
We want to be able to have more than one tear on its face at once.
First step is to change how we're remembering whether there's a tear.
We need to keep track of a list --- see

https://docs.python.org/3.4/tutorial/introduction.html#lists

for details --- of the y-values of all the tears currently on the face.
At the moment there can only be one, though.